backend/requirements.txt
Flask
Flask-Cors
python-docx
fpdf
reportlab
